https://issues.apache.org/jira/browse/AMQCLI-2
Adding the first version of JAXB classes generated from the Artemis
import schema.  There is also a marshaller class and unit test to
demonstrate streaming to an XML file when marshalling instead of writing
the entire file at once. This will need some work but is a good starting
point.

Also Artemis is currently a SNAPSHOT dependency as we need to work off
the newest code in master.